requirements.txt
oauthlib==0.6.2
python-twitter==1.3.1
requests==2.3.0
requests-oauthlib==0.4.1
simplejson==3.5.2
